refactor(container): mv CommunitySetter to tool scope
