+++ src/containers/tool/CommunitySetter/styles/search_bar.js
