+++ src/containers/tool/CommunitySetter/tests/store.test.js
